Check for neutral lines explicitly
This seems more robust (given the information known about the unified
diff format) than the search for end lines used previously. One edge
case that remains is the handling of hunk lines that indicate there was
no newline at the end of a file. These begin with a leading \ rather
than a leading space, +, or -.